requirements.txt
chromadb==0.6.3
langchain-community==0.3.16
transformers==4.47.0
torch==2.5.1
pandas>=2.2.2
numpy>=1.26.4
pydantic>=2.10.3
ragas==0.2.12
sentence-transformers==3.3.1
google-generativeai==0.8.3
langchain-google-genai==2.0.9
langchain==0.3.17
langchain-core==0.3.33
datasets>=3.2.0
xmltodict==0.14.2
jsonpath==0.82.2
scikit-learn>=1.2.2
pyyaml>=6.0.2
httpx>=0.23.0
tqdm>=4.67.1
typing-extensions>=4.12.2